requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
blinker==1.9.0
boto3==1.41.0
botocore==1.41.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.2
crispy-bootstrap5==2025.4
cryptography==46.0.2
cycler==0.12.1
distlib==0.3.9
Django==5.0.2
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
django-humanize==0.1.2
django-js-asset==3.1.2
django-phonenumber-field==8.1.0
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
dotenv==0.9.9
fastapi==0.116.1
filelock==3.16.1
Flask==3.1.0
fonttools==4.59.0
gunicorn==23.0.0
h11==0.16.0
humanize==4.12.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mysql-connector-python==9.0.0
mysqlclient==2.2.7
num2words==0.5.14
numpy==2.3.1
packaging==25.0
pandas==2.3.1
phonenumbers==9.0.4
pillow==11.3.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
qrcode==7.4.2
razorpay==1.4.1
requests==2.32.3
s3transfer==0.14.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.27.0
virtualenvwrapper-win==1.2.7
Werkzeug==3.1.3
